move test cert generation to interceptor_suite.go
